+++ smqproxy-20260409/alibabacloud_smqproxy20260409/models/_receive_message_response_body.py
@@ -2,6 +2,9 @@
 # This file is auto-generated, don't edit it. Thanks.
 from __future__ import annotations
 
+from typing import Dict
+
+from alibabacloud_smqproxy20260409 import models as main_models
 from darabonba.model import DaraModel
 
 class ReceiveMessageResponseBody(DaraModel):
@@ -17,7 +20,7 @@ class ReceiveMessageResponseBody(DaraModel):
         next_visible_time: int = None,
         priority: int = None,
         receipt_handle: str = None,
-        user_properties: str = None,
+        user_properties: Dict[str, main_models.UserPropertiesValue] = None,
     ):
         self.dequeue_count = dequeue_count
         self.enqueue_time = enqueue_time
@@ -32,7 +35,10 @@ class ReceiveMessageResponseBody(DaraModel):
         self.user_properties = user_properties
 
     def validate(self):
-        pass
+        if self.user_properties:
+            for v1 in self.user_properties.values():
+                 if v1:
+                    v1.validate()
 
     def to_map(self):
         result = dict()
@@ -69,8 +75,10 @@ class ReceiveMessageResponseBody(DaraModel):
         if self.receipt_handle is not None:
             result['ReceiptHandle'] = self.receipt_handle
 
+        result['UserProperties'] = {}
         if self.user_properties is not None:
-            result['UserProperties'] = self.user_properties
+            for k1, v1 in self.user_properties.items():
+                result['UserProperties'][k1] = v1.to_map() if v1 else None
 
         return result
 
@@ -106,8 +114,11 @@ class ReceiveMessageResponseBody(DaraModel):
         if m.get('ReceiptHandle') is not None:
             self.receipt_handle = m.get('ReceiptHandle')
 
+        self.user_properties = {}
         if m.get('UserProperties') is not None:
-            self.user_properties = m.get('UserProperties')
+            for k1, v1 in m.get('UserProperties').items():
+                temp_model = main_models.UserPropertiesValue()
+                self.user_properties[k1] = temp_model.from_map(v1)
 
         return self